--- constants/EXAMPLE_CONST.js
@@ -1,5 +1,0 @@
-const EXAMPLE_CONST = {
-  example: "example",
-};
-
-export default EXAMPLE_CONST;
