+++ controller/recipes/deleteFromShoppingList.js
@@ -1,10 +1,34 @@
+import User from "../../service/schemas/user.js";
+
 const deleteFromShoppingList = async (req, res, next) => {
+  const userId = req.user._id;
+  const { ingredientId } = req.body;
+
   try {
-    // code
-    res.json({ message: "OK" });
-  } catch (error) {
-    console.log(error);
-    next(error);
+    const user = await User.findById(userId);
+
+    if (!ingredientId) {
+      return res.status(400).json({ message: "Ingredient ID is required" });
+    }
+
+    const existingItemIndex = user.shoppingList.findIndex(
+      (item) => item.ingredient._id.toString() === ingredientId.toString(),
+    );
+
+    if (existingItemIndex === -1) {
+      return res
+        .status(404)
+        .json({ message: "Ingredient not found in your shopping list" });
+    }
+
+    user.shoppingList = user.shoppingList.filter(
+      (item) => item.ingredient.toString() !== ingredientId.toString(),
+    );
+
+    await user.save();
+    res.status(200).json({ message: "Ingredient removed from shopping list" });
+  } catch (err) {
+    next(err);
   }
 };
