+++ controller/recipes/getShoppingList.js
@@ -1,10 +1,36 @@
-const getShoppingList = = async (req, res, next) => {
+const getShoppingList = async (req, res, next) => {
+  const userId = req.user._id;
+  const { ingredientId, quantity, measure } = req.body;
+
   try {
-    // code
-    res.json({ message: "OK" });
-  } catch (error) {
-    console.log(error);
-    next(error);
+    const user = await User.findById(userId).populate(
+      "shoppingList.ingredient",
+    );
+    const ingredient = await Ingredient.findById(ingredientId);
+
+    if (!ingredient) {
+      return res.status(404).json({ message: "Ingredient not found" });
+    }
+
+    const existingItemIndex = user.shoppingList.findIndex(
+      (item) => item.ingredient._id.toString() === ingredientId.toString(),
+    );
+
+    if (existingItemIndex > -1) {
+      user.shoppingList[existingItemIndex].quantity += quantity;
+    } else {
+      user.shoppingList.push({ ingredient: ingredientId, quantity, measure });
+    }
+
+    await user.save();
+
+    const updatedUser = await User.findById(userId).populate(
+      "shoppingList.ingredient",
+    );
+
+    res.status(200).json(updatedUser.shoppingList);
+  } catch (err) {
+    next(err);
   }
 };
