+++ routes/others.js
@@ -2,14 +2,14 @@
 import express from "express";
 
 // my modules
+import addIngredientToList from "../controller/recipes/addToShoppingList.js";
+import removeIngredientFromList from "../controller/recipes/deleteFromShoppingList.js";
 import getMyFavoriteRecipes from "../controller/recipes/getMyFavoriteRecipes.js";
+import getIngredientList from "../controller/recipes/getShoppingList.js";
 import searchRecipeByKeyword from "../controller/recipes/searchRecipeByKeyword.js";
 import setMyFavoriteRecipe from "../controller/recipes/setMyFavoriteRecipe.js";
 import subscribeNewsletter from "../controller/recipes/subscribeNewsletter.js";
 import unsetMyFavoriteRecipe from "../controller/recipes/unsetMyFavoriteRecipe.js";
-import addIngredientToList from "../controller/shopping-list/addIngredientToList.js";
-import getIngredientList from "../controller/shopping-list/getIngredientList.js";
-import removeIngredientFromList from "../controller/shopping-list/removeIngredientFromList.js";
 import authorization from "../middlewares/jwtMiddleware.js";
 
 const route = express.Router();
