--- routes/search.js
@@ -1,11 +1,0 @@
-// //npm modules
-// import express from "express";
-
-// //my modules
-// import authorization from "../middlewares/jwtMiddleware.js";
-
-// const route = express.Router();
-
-// route.use(authorization);
-
-// export default route;
